Create Verification
Working project file. Logic is working with GET requests but I have not figured out or implemented any authentication.
Using a workaround by opening the URLs in a browser that is logged into the system I need to check.
Will make more usable in the future, eg adding menu, etc.